big refractor of the scraper
