+++ src/helpers/delay.ts
@@ -1,0 +1,5 @@
+export function delay(time: number) {
+    return new Promise(function(resolve) { 
+        setTimeout(resolve, time)
+    });
+}
